backend/requirements.txt
django
gunicorn==20.1.0
psycopg2-binary==2.9.2
django-environ==0.8.1
python-dotenv
djangorestframework
markdown
django-filter
Pillow
dj-rest-auth
django-allauth
django-cors-headers
pycodestyle==2.8.0
python-dateutil==2.8.2